Add app's default look-n-fell
In 1aec494 the bootstrap default look-n-feel was slightly modified
to comply with UCR's standard colour pallete. Here, the temptative
design of the main's navbar of Tydilabs is added, using the custom
colours.

The needed controller also was created and with it, all the genera-
ted files created by rails scaffold.