Fix and redesign navbar
Remove 'navbar-banner' and implement standard bootstrap navbar for simplicity

'navbar-banner' was intended to fit two logos but it was way too big and
didn't work well in mobile. A standard bootstrap navbar with
'navbar-brand' to fit a logo looks like a better idea.

However only one logo is displayed. The logo displayed is 'belt.png'
(client office) which is more relevant to the app than the university
logo.

Maybe the university logo can be added in grey-scale to the footer
later on.