Add helper method for asset type
Add method that converts the asset type to a human readble name.
Asset type returns the submodel name (ex. PlatedAsset) and in the views
the asset type is often required, so the asset type name conversion
needed to be DRY.

asset#type_to_h convertes the submodel name to a friendly name (ex.
PlatedAsset -> Placa)