Add identification code feature
Adds asset identification code feature. Allows an asset to be uniquely
identified by scanners via QR code.

The id code is generated for each asset after it's created. For existing
assets use 'rake db_record_updater:generate_id_codes'.

Also the QR code is generated and rendered in views 'on the fly' to
avoid dealing with files and delayed jobs (may change in the future).
For now the id code can't be re-generated or deleted.

ALL assets will have an id code and will render the QR code in views.