Improve functionality of AssetCategory
Now the category of an asset will defined whether an
asset can have details such as warranty, technical,
network or security. This avoids showing an info
message telling the user to add such details to
assets where those details don't apply.

The AssetCategory crud was improved, the screens to
show and update a given asset were added. The create
form was refactored to work as both, an edit and
create form.