Add print feature for identification codes
Identification codes (QR) can be printed needed to be printed in order
to identfy assets via scanners or tidylabs mobile. The codes can be
printed either individually or in a group.

Group printing allows id codes located in the SAME area to be printed
in a single batch, reducing paper consumption and labelling an entire
area easier.

Individual printing gives more flexibility for other scenarios, like
when a new asset is registered or the reprint of code is needed.