Make app title customizable
Allows the app title to be customized from any view.
Use ```<% title 'Your title here' %>``` in the view to set the custom
title.

When a custom title isn't specified in the view, the default title is
shown: "Oficina De Computo - UCR | Tydilabs".

When a title is specified, it will be changed to:
"Oficina De Computo - UCR | <your title here>"